+++ modules/bucket/outputs.tf
@@ -1,9 +1,9 @@
 output "bucket_name" {
   description = "The name of the bucket"
-  value       = var.bucket != null ? google_storage_bucket.this[0].name : null
+  value       = var.bucket_name_prefix != "" ? google_storage_bucket.tf_states[0].name : null
 }
 
 output "bucket_labels" {
   description = "The labels of the bucket"
-  value       = var.bucket != null ? google_storage_bucket.this[0].labels : null
+  value       = var.bucket_name_prefix != "" ? google_storage_bucket.tf_states[0].labels : null
 }