+++ modules/vpc/outputs.tf
@@ -1,9 +1,9 @@
 output "vpc_name" {
   description = "Name of the project vpc"
-  value       = var.create_vpc ? google_compute_network.vpc[0].name : null
+  value       = var.vpc_create ? google_compute_network.vpc[0].name : null
 }
 
 output "vpc_self_link" {
   description = "Self link to the project vpc"
-  value       = var.create_vpc ? google_compute_network.vpc[0].self_link : null
+  value       = var.vpc_create ? google_compute_network.vpc[0].self_link : null
 }