+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.0
+1.0.0 is a major backwards incompatible release
+### ADDED
+- Remove default service account from editors 
+- Add README
+### CHANGED
+- Create only one default service account with default roles
+- Each togglable feature of the module enables necesasry gcloud apis and assigns necessary roles to default sa
+- Create only one bucket if enabled for terraform states
+- Allow multiple providers corresponding to different repos to be configured for the github identity pool if enabled
+
 ## 0.3.0
 ### ADDED
 - Allow gha provider partially match repo based on `gha_wif_org` when `gha_wif_repo` is an empty string.